requirements.txt
pbr>=5.7.0
qiskit-terra>=0.24.0
